sam3-api/requirements.txt
fastapi==0.115.0



uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow>=10.0.0
numpy>=1.26,<2
opencv-python-headless>=4.8,<5

torch
torchvision
ultralytics>=8.4.21
timm>=1.0.17
tqdm
ftfy==6.1.1
regex
iopath>=0.1.10
typing_extensions
huggingface_hub
einops
psutil
